made changes to service and controller for soft delete

+++ src/controllers/url.controller.js
@@ -6,7 +6,7 @@ export class UrlController {
 
     if (!original_url) {
       return res.status(400).json({
-        error: "No original url could be found",
+        error: "No original URL provided",
       });
     }
 
@@ -44,15 +44,15 @@ export class UrlController {
 
     if (!code) {
       return res.status(400).json({
-        error: "No short code could be found",
+        error: "No short code provided",
       });
     }
 
     try {
       const url = await UrlService.findByShortCode(code);
 
-      if (!url) {
-        return res.status(404).send("No Original Urls could be found");
+      if (!url || url.deletedAt) {
+        return res.status(404).send("No original URL found");
       }
 
       await UrlService.updateClickCount(code);
@@ -70,13 +70,40 @@ export class UrlController {
 
     if (!code) {
       return res.status(400).json({
-        error: "No short code could be found",
+        error: "No short code provided",
       });
     }
 
+    let apiKey =
+      req.headers["api-key"] ||
+      req.headers["x-api-key"] ||
+      req.body.apiKey ||
+      req.headers.authorization;
+
+    if (apiKey?.startsWith("Bearer ")) {
+      apiKey = apiKey.split(" ")[1];
+    }
+
     try {
+      let user = null;
+      if (apiKey !== undefined) {
+        user = await UrlService.findUserByApiKey(apiKey);
+      }
+
+      const url = await UrlService.findByShortCode(code);
+
+      if (!url || url.deletedAt) {
+        return res.status(404).send("No original URL found");
+      }
+
+      if (user && url.userId !== user.id) {
+        return res.status(403).json({
+          error: "You do not have permission to delete this URL",
+        });
+      }
+
       await UrlService.delete(code);
-      return res.status(204).send("Url deleted successfully");
+      return res.status(204).send("URL deleted successfully");
     } catch (error) {
       return res.status(500).json({
         error: error.message,
